requirements.txt
slack-bolt==1.21.3
slack-sdk==3.34.0
chromadb==0.6.3
ollama==0.4.7
httpx==0.28.1
python-dotenv==1.1.0
